requirements.txt
amqp==1.4.9
ansible==2.7.10
anyjson==0.3.3
asn1crypto==0.24.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.23
celery==3.1.25
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
configobj==4.7.2
cryptography==2.2.2
decorator==3.4.0
Django==1.11.23
django-bootstrap-form==3.2
django-celery==3.1.17
django-cors-headers==1.1.0
django-redis==4.10.0
django-redis-cache==2.0.0
djangorestframework==3.9.4
ecdsa==0.13
enum34==1.1.6
idna==2.8
iniparse==0.4
ipaddress==1.0.16
IPy==0.75
Jinja2==2.7.3
kombu==3.0.37
MarkupSafe==0.9.3
meld3==1.0.2
MySQL-python==1.2.5
paramiko==1.17.4
perf==0.1
pip2pi==0.7.0
policycoreutils-default-encoding==0.1
pycparser==2.19
pycrypto==2.6.1
pycurl==7.19.0
pygobject==3.22.0
pygpgme==0.3
pyinotify==0.9.6
pyliblzma==0.5.3
pymssql==2.1.4
python-linux-procfs==0.4.9
pytoml==0.1.14
pytz==2016.3
pyudev==0.15
pyvmomi==6.0.0.2016.4
pyxattr==0.5.1
PyYAML==3.10
redis==2.10.6
registries==0.1
requests==2.22.0
schedutils==0.4
seobject==0.1
sepolicy==1.1
singledispatch==3.4.0.3
six==1.10.0
slip==0.4.0
slip.dbus==0.4.0
supervisor==3.3.1
tornado==4.3
urlgrabber==3.10
urllib3==1.22
vine==1.3.0
XlsxWriter==0.9.3
xlwt==1.1.2
yum-metadata-parser==1.1.4
